+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Version 0.1.1 (June 21, 2022)
+- **Features**
+  - Add Editor tile spawning button
+
+- **Bug Fixes**
+  - Add authorized tree positions based on OSM Map
+
 ## Version 0.1.0 (May 31, 2022)
 - **Features**
   - Intial release